pyproject.toml
[project]
name = "revelai-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Ivisson Alves"}
]
license = {text = "MIT"}
requires-python = ">=3.12, <4.0.0"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "pymongo[srv] (>=4.16.0,<5.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "langchain-groq (>=1.1.2,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "certifi (>=2026.2.25,<2027.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
